phase8(gate0): uint casts for DXGI/texture descriptions (box compile)
The x64 box's first compile flagged the five int->uint sites in
ResizeBuffers args and the SwapChainDescription1/Texture2DDescription
Width/Height fields — the risk the Gate 0 report ranked #2. The WMC9999
XAML error was the usual MarkupCompilePass2 cascade (WMC1509) and
disappeared with the CS fixes.

Gate 0 verified on the x64 box (RX 9070 XT, interactive session via
schtasks): soak3d: cycles=50 created=50 released=50 removed=0 errors=0.
SetSwapChain bound at the panel's real 440x225; create-on-show /
release-on-hide paired cleanly every cycle.

+++ src/TokenBar.App/Graph3DRenderer.cs
@@ -174,7 +174,7 @@ internal sealed class Graph3DRenderer : IDisposable
         _depthView.Dispose();
         _depth.Dispose();
         _swapChain.ResizeBuffers(
-            2, width, height, Format.B8G8R8A8_UNorm, SwapChainFlags.None);
+            2, (uint)width, (uint)height, Format.B8G8R8A8_UNorm, SwapChainFlags.None);
         _width = width;
         _height = height;
         CreateBackbufferViews(width, height);
@@ -208,8 +208,8 @@ internal sealed class Graph3DRenderer : IDisposable
         // SwapChainPanel's DirectComposition visual owns presentation.
         var desc = new SwapChainDescription1
         {
-            Width = width,
-            Height = height,
+            Width = (uint)width,
+            Height = (uint)height,
             Format = Format.B8G8R8A8_UNorm,
             Stereo = false,
             SampleDescription = new SampleDescription(1, 0),
@@ -256,8 +256,8 @@ internal sealed class Graph3DRenderer : IDisposable
         _rtv = _device.CreateRenderTargetView(backbuffer);
         _depth = _device.CreateTexture2D(new Texture2DDescription
         {
-            Width = width,
-            Height = height,
+            Width = (uint)width,
+            Height = (uint)height,
             MipLevels = 1,
             ArraySize = 1,
             Format = Format.D32_Float,
